requirements.txt
absl-py==2.3.1
attrs==25.3.0
cattrs==1.5.0
cloudpickle==3.1.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.0
grpcio==1.74.0
gym==0.26.2
gym-notices==0.1.0
h5py==3.14.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
kiwisolver==1.4.7
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.9.4
mlagents==0.30.0
mlagents-envs==0.30.0
numpy==1.21.2
packaging==25.0
PettingZoo==1.15.0
pillow==11.3.0
protobuf==6.32.0
pygame-ce==2.5.6
pygame_gui==0.6.13
pyparsing==3.2.4
pypiwin32==223
python-dateutil==2.9.0.post0
python-i18n==0.3.9
pywin32==311
PyYAML==6.0.2
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
Werkzeug==3.1.3
zipp==3.23.0
